requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
requests>=2.32.0
google-generativeai>=0.8.0
tavily-python>=0.5.0
pydantic-settings>=2.6.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
